// frontend/src/api/generated/skills/skills.ts
export type searchClawhubApiV1SkillsClawhubSearchGetResponse200 = {
  data: ClawHubSearchResult[];
  status: 200;
};

export type searchClawhubApiV1SkillsClawhubSearchGetResponse422 = {
  data: HTTPValidationError;
  status: 422;
};

export type searchClawhubApiV1SkillsClawhubSearchGetResponseSuccess =
  searchClawhubApiV1SkillsClawhubSearchGetResponse200 & {
    headers: Headers;
  };
export type searchClawhubApiV1SkillsClawhubSearchGetResponseError =
  searchClawhubApiV1SkillsClawhubSearchGetResponse422 & {
    headers: Headers;
  };

export type searchClawhubApiV1SkillsClawhubSearchGetResponse =
  | searchClawhubApiV1SkillsClawhubSearchGetResponseSuccess
  | searchClawhubApiV1SkillsClawhubSearchGetResponseError;